Don't show Contextual Forms if registry has no contexts.
Fixes bug #TRRFFDA-1

The bug manifested itself when the registry had a context form group
defined, but the "context" feature wasn't enabled on the registry.

The contextual form group still showed up with no forms, but an Add
button, that produced a 404 when clicked.

After the fix we will not show the contextual form groups if the registry doesn't
have the "contexts" feature.