#846 fix the reset password link
Note there are two issues:
* first the class UserChangeForm is overriding the base Django UserChangeForm class
in order to add support for workinggroups and registry fields. If at the time of the
decision it was ok, since Django likely updated the Django UserChangeForm and so we
are losing the new code from Django. Probably introducing other bug (.i.e some code is
related to user permission in the Django UserChangeForm). An alternative solution is likely to extend
the base function instead to override it.
* the password reset page has the default Django UI theme - I guess it has always been the
case.